+++ lib/src/features/movies/views/movie_list_page.dart
@@ -4,8 +4,8 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:infinite_scroll_pagination/infinite_scroll_pagination.dart';
 import 'package:tmdb_app/src/features/movies/controller/movie_controller.dart';
 import 'package:tmdb_app/src/features/movies/data_model/movie_response/movie/movie.dart';
-import 'package:tmdb_app/src/features/movies/views/component/movie_list.dart';
-import 'package:tmdb_app/src/features/movies/views/component/upcoming_movie_list.dart';
+import 'package:tmdb_app/src/features/movies/views/part/movie_list.dart';
+import 'package:tmdb_app/src/features/movies/views/part/upcoming_movie_list.dart';
 
 class MovieListPage extends HookConsumerWidget {
   const MovieListPage({super.key});
