+++ lib/src/features/movies/views/part/review_list.dart
@@ -3,10 +3,10 @@ import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:flutter_staggered_animations/flutter_staggered_animations.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:infinite_scroll_pagination/infinite_scroll_pagination.dart';
-import 'package:tmdb_app/src/common_widgets/cached_image.dart';
 import 'package:tmdb_app/src/features/movies/controller/movie_controller.dart';
 import 'package:tmdb_app/src/features/movies/data_model/review_response/review/review.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:tmdb_app/src/features/movies/views/component/review_card.dart';
 
 class ReviewList extends HookConsumerWidget {
   final String movieId;
@@ -71,25 +71,7 @@ class ReviewList extends HookConsumerWidget {
               child: SlideAnimation(
                 verticalOffset: 50.0,
                 child: FadeInAnimation(
-                  child: ListTile(
-                    leading: CachedImage(
-                      imageURL: item.authorDetails["avatar_path"] != null
-                          ? "https://image.tmdb.org/t/p/w500${item.authorDetails["avatar_path"]}"
-                          : "assets/images/cinema.jpeg",
-                      width: 50,
-                      height: 50,
-                      isCircle: true,
-                    ),
-                    title: Text(
-                      item.author,
-                      style: Theme.of(context).textTheme.bodyLarge,
-                    ),
-                    subtitle: Text(
-                      item.content,
-                      style: Theme.of(context).textTheme.bodySmall,
-                      overflow: TextOverflow.ellipsis,
-                    ),
-                  ),
+                  child: ReviewCard(item: item),
                 ),
               ),
             );
